refactor: split app schedule editor helpers
